Update README.md
How does Chatbot Works?
Chatbots are computer programs that simulate conversation with humans. They’re used in a variety of applications, from providing customer service to answering questions on a website.

Here’s a general breakdown of how a chatbot works:

User Input: The user starts a conversation with the chatbot by typing in a message or speaking to it through a voice interface.
Understanding the User: The chatbot analyzes the user’s input using NLP. For rule-based chatbots, this involves matching keywords and phrases. For AI-powered chatbots, it’s more complex and involves understanding the intent behind the user’s words.
Generating a Response: Based on its understanding of the user’s input, the chatbot retrieves a response from its database. This response could be a simple answer, a more complex explanation, or even a question to clarify the user’s intent.
Conversation Flow: The chatbot delivers the response to the user, and the conversation continues. The user can provide additional information or ask follow-up questions, and the chatbot will respond accordingly.
Building A Chatbot Using Python
Now we have an immense understanding of the theory of chatbots and their advancement in the future. Let’s make our hands dirty by building one simple rule-based chatbot using Python for ourselves.